fixed bug where goal state had to be set to true
there were a few bugs here.

1) getHeuristic was returning a large number for getHeuristic(GOAL_STATE)
instead of 0
2) pathpostprocessor still isn't really doing the right thing for applying the
mprim to the second to last state. because the resulting successor will have a
non-GOAL_STATE id, you have to ensure the successor satisfies the goal
constraints. i was lazy, so i just said "if it's equal to 1, it's good enough".
need to fix this.